WARNING: MATCH AND RECONSTRUCTION DATA BREAK!!!
Changing all disk I/O to be done with the Cereal Library. This will make all
former match and reconstruction files incompatible with the cereal versions. The
API does not change for any read/write methods, but the binary formats of the
data will be different and so old files can no longer be read in this way.

Change-Id: I95a23b7d6be6b3e8115eb6049b4f9e9241ccdab8