Large refactor of bundle adjustment and track estimation.
This commit was originally split into 3 but I had some git troubles... sorry for
the monster commit!

All BA now runs through the partial BA. This removes duplicate code and allows
for a consistent interface to optimize camera intrinsics. The track estimator
was updated to be more clean internally with a better interface.

Change-Id: I64cc58532b05ecf8ba17d35238345740f7522e09