Slight adjustment of triangulation outlier removal.
Previously, we filtered (i.e. removed) 3D points that had at least 1 observation
with a reprojection error greater than a certain threshold. With the new ability
to use robust cost functions for bundle adjustment, it makes more sense to
filter based on the mean of the reprojection errors since BA can still be robust
to a single outlier.

Change-Id: I02118a8e4122963ef077ccfa726f1e730ab84274