Add Camera intrinsic groups to Reconstruction class thanks to @stoyanovd
When adding a view to a reconstruction there is now additional ability to set
which camera intrinsic group it belongs to. This is useful for datasets where a
single camera captures multiple images, and so the camera intrinsics should be
optimized jointly as the same intrinsics for all of those images. The interface
is through the Reconstruction class so to provide safe access and management of
the data.

Change-Id: I93bcba145db3ada85e87874c471c88e5942d0ffa