Refactor reprojection error to allow generic camera models.
The reprojection error cost function is now templated so that various camera
model classes can be passed to it as the template parameter. This allows the
reprojection error method to directly call the camera intrinsics model class to
perform camera projection (through the template parameter) and simplifies the
code in a way that is extendible going forward.

Change-Id: I98f31c04af8f6cc4296b76cea916163e8f78510d