Resolution-aware thresholds for 2-view verification and pose estimation.
The 2-view geometric verification and the absolute pose module for incremental
sfm now use thresholds that scale with the resolution of the image. This is so
that a single threshold may be used in a dataset with images of varying
resolution. This is done in the original PhotoTourism paper where they describe
thresholds in terms of the % of the maximum image dimension instead of pixels.

Change-Id: I79e91890bd9486fb1bc504b1a08f08d699403c21