Optimization of shared camera intrinsics.
The previous "shared camera intrinsics" mechanism was somewhat hacked together
to preserve the previous structure of the Camera class. Now, we use an explicit
shared_ptr inside the Camera class to maintain intrinsics that are shared
between potentially many cameras. This change *explicitly* shares the memory and
associated data for the shared intrinsics. This simplifies the logic in a few
modules, notably bundle adjustment where we can remove a few conditionals on the
inner loop which increases performance noticeably.

Change-Id: Iea57be54b02d2facd0dd5aa3ac51eaa548b569e8