Track subset selection for efficient bundle adjustment.
Bundle adjustment often contains many more points than is strictly necessary in
order to converge to a high quality solution. This new method allows for tracks
to be subsampled so that only the highest quality tracks are chosen for BA. The
tracks are chosen in a way such that views are well-constrained.

Change-Id: I82363bdb91239cdec829348a4a2ad9f6401c73dd