MegaDiff - Refactor of feature extraction and matching.
The logic for storing feature descriptors and matches has been
moved out of FeatureMatcher and into a FeaturesAndMatchesDatabase.
This database provides an abstract interface for adding and removing
the features and match information which is typically heavy in memory.
The previous "in core" and "out of core" settings are refactored as
"in-memory" and "local" implementations of the
FeaturesAndMatchesDatabase.

This also allowed a refactor of FeatureMatcher to simplify the interface
by reusing the FeaturesAndMatchesDatabase. The global descriptor
matching was moved to the FeatureExtractorAndMatcher to reduce code bloat.

This refactor breaks the API for the matching.